pyproject.toml
[project]
name = "deep-research-swarm"
version = "0.1.0"
description = "Multi-agent deep research system with STORM-style decomposition and RAG-Fusion synthesis"
requires-python = ">=3.11"
dependencies = [
    # Execution
    "langgraph>=1.0.9,<2.0",
    "langgraph-checkpoint-sqlite>=3.0.3,<4.0",
    # LLM
    "anthropic>=0.49.0,<1.0",
    # Search
    "exa-py>=2.4.0,<3.0",
    "tavily-python>=0.7.21,<1.0",
    "httpx>=0.27.0,<1.0",
    # Extraction
    "crawl4ai>=0.8.0,<1.0",
    "trafilatura>=2.0.0,<3.0",
    "pymupdf4llm>=0.3.4,<1.0",
    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-cov", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["deep_research_swarm*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
